kiosk_backend/requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==10.4.0
django-filter==24.3
python-dateutil==2.9.0.post0
reportlab==4.2.5
python-dotenv==1.0.1
drf-spectacular==0.27.2
pyserial>=3.5
pythonnet>=3.0.0
openpyxl==3.1.2
jdatetime==4.1.0
python-escpos>=3.0
gunicorn==23.0.0
waitress>=3.0.0
requests>=2.31.0
pywebview>=5.0.0
nuitka>=2.0.0

